# api/openapi.yaml
openapi: 3.0.3
info:
  title: LequePlay API
  description: |
    API da **LequePlay** — plataforma social para descobrir, acompanhar e avaliar
    filmes, séries e podcasts.

    Esta entrega cobre o **catálogo público** e a **busca semântica** (embeddings
    Gemini + pgvector, feita no backend) e o **login via Keycloak** (proxy de
    password grant + validação de JWT).

    ## Autenticação
    Endpoints protegidos exigem um **Bearer token** (JWT do realm `lequeplay`),
    obtido em `POST /v1/auth/login`. No cabeçalho: `Authorization: Bearer <token>`.
    Endpoints do catálogo e da busca são **públicos**.

    ## Erros
    Erros seguem o formato `application/problem+json` (RFC 7807).
  version: 0.1.0
  contact:
    name: Um Leque de Tecnologia
  license:
    name: Proprietary
servers:
  - url: https://api.lequeplay.rodolfodebonis.com.br
    description: Produção
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /v1/midias:
    get:
      summary: Lista o catálogo
      description: Lista mídias com filtros, ordenadas por popularidade, paginadas.
      tags: [Catálogo]
      security: []
      parameters:
        - $ref: '#/components/parameters/Tipo'
        - $ref: '#/components/parameters/Genero'
        - $ref: '#/components/parameters/Ano'
        - $ref: '#/components/parameters/Q'
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/PorPagina'
        - $ref: '#/components/parameters/Limite'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Página de mídias
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginaMidias' }
  /v1/midias/{id}:
    get:
      summary: Detalhe de uma mídia
      description: >-
        Retorna a mídia com créditos (elenco/equipe) e temporadas (séries).
        O parâmetro aceita o id (UUID) ou o slug da mídia.
      tags: [Catálogo]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: id (UUID) ou slug da mídia, ex. `homem-aranha-um-novo-dia`.
          schema: { type: string }
      responses:
        '200':
          description: Detalhe da mídia
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MidiaDetalhe' }
        '404':
          $ref: '#/components/responses/NaoEncontrado'
  /v1/midias/{id}/temporadas/{numero}/episodios:
    get:
      summary: Episódios de uma temporada
      description: >-
        Lista os episódios de uma temporada de uma série. A mídia é referenciada
        pelo id (UUID) ou pelo slug; a temporada, pelo número.
      tags: [Catálogo]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: id (UUID) ou slug da mídia (série).
          schema: { type: string }
        - name: numero
          in: path
          required: true
          description: Número da temporada.
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Episódios da temporada (lista vazia se ainda não catalogados)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EpisodioTemporada' }
        '400':
          $ref: '#/components/responses/Validacao'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
  /v1/busca:
    get:
      summary: Busca no catálogo
      description: |
        Busca por relevância. Modos:
        - `auto` (padrão) / `hybrid`: fusão RRF de busca semântica (vetorial) + léxica.
        - `vector`: puramente semântica (embeddings Gemini + pgvector, cosseno).
        - `fts`: puramente léxica (full-text português + trigram).

        Quando o Gemini está indisponível, cai para `fts` e marca `usouFallback`.
      tags: [Busca]
      security: []
      parameters:
        - name: q
          in: query
          required: true
          description: Consulta em linguagem natural.
          schema: { type: string, example: crime e investigação policial }
        - name: modo
          in: query
          required: false
          schema:
            type: string
            enum: [auto, hybrid, vector, fts]
            default: auto
        - $ref: '#/components/parameters/Tipo'
        - $ref: '#/components/parameters/Genero'
        - $ref: '#/components/parameters/Ano'
        - $ref: '#/components/parameters/Limite'
      responses:
        '200':
          description: Resultado da busca
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResultadoBusca' }
        '400':
          $ref: '#/components/responses/Validacao'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT do realm Keycloak `lequeplay` (obtido em /v1/auth/login).
  parameters:
    Tipo:
      name: tipo
      in: query
      required: false
      schema: { type: string, enum: [filme, serie, podcast] }
    Genero:
      name: genero
      in: query
      required: false
      schema: { type: string, example: Drama }
    Ano:
      name: ano
      in: query
      required: false
      schema: { type: integer, example: 2022 }
    Limite:
      name: limite
      in: query
      required: false
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    Offset:
      name: offset
      in: query
      required: false
      schema: { type: integer, default: 0, minimum: 0 }
    Pagina:
      name: pagina
      in: query
      required: false
      description: 'Página, começando em 1. Quando vem junto com offset, este par vence.'
      schema: { type: integer, default: 1, minimum: 1 }
    PorPagina:
      name: porPagina
      in: query
      required: false
      description: 'Itens por página. Teto de 100, como em limite.'
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    Q:
      name: q
      in: query
      required: false
      description: >-
        Filtro textual simples da listagem: casa com título ou sinopse. Para a
        busca que entende intenção — vetorial e híbrida — use /v1/busca.
      schema: { type: string, example: carbono }
  responses:
    Validacao:
      description: Requisição inválida
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    NaoEncontrado:
      description: Recurso não encontrado
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Midia:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string, description: 'Identificador legível e estável, aceito em /v1/midias/{id}.' }
        tipo: { type: string, enum: [filme, serie, podcast] }
        titulo: { type: string }
        tituloOriginal: { type: string }
        sinopse: { type: string }
        ano: { type: integer }
        generos:
          type: array
          items: { type: string }
        posterUrl: { type: string, format: uri }
        duracaoMin: { type: integer }
        popularidade: { type: number, format: float }
        notaMedia: { type: number, format: float }
        totalAvaliacoes:
          type: integer
          description: >-
            Quantas pessoas avaliaram. Vem sempre, inclusive como 0 — é ele, e
            não notaMedia, que responde "este título tem nota?". Sem o contador,
            um título que ninguém avaliou fica indistinguível de um que todo
            mundo detestou.
        status:
          type: string
          description: >-
            Situação de produção, como a TMDB publica: "Returning Series",
            "Ended", "Canceled" para série; "Released" para filme. Ausente
            quando a origem não informa.
          example: Ended
        frequencia:
          type: string
          description: 'Periodicidade do podcast (ex.: Semanal). Apenas para tipo=podcast; ausente nos demais.'
          example: Semanal
      required: [id, slug, tipo, titulo, generos, popularidade, notaMedia, totalAvaliacoes]
    MidiaDetalhe:
      allOf:
        - $ref: '#/components/schemas/Midia'
        - type: object
          properties:
            creditos:
              type: array
              items: { $ref: '#/components/schemas/Credito' }
            temporadas:
              type: array
              items: { $ref: '#/components/schemas/Temporada' }
            episodios:
              type: array
              description: Episódios do podcast (apenas para tipo=podcast).
              items: { $ref: '#/components/schemas/Episodio' }
    PessoaResumo:
      type: object
      description: >-
        Quem assina um crédito. Objeto, e não o nome solto: nome não é
        identificador (existem dois "João Silva"), e a tela da pessoa precisa de
        um endereço estável. O slug segue a mesma regra do slug da mídia.
      properties:
        slug: { type: string, example: marina-aleixo }
        nome: { type: string, example: Marina Aleixo }
        fotoUrl: { type: string, format: uri }
      required: [slug, nome]
    Credito:
      type: object
      properties:
        pessoa: { $ref: '#/components/schemas/PessoaResumo' }
        papel: { type: string, enum: [direcao, elenco, apresentacao] }
        personagem: { type: string, description: 'Só vem preenchido quando papel = elenco.' }
      required: [pessoa, papel]
    Temporada:
      type: object
      properties:
        numero: { type: integer }
        nome: { type: string }
        ano: { type: integer }
        totalEpisodios: { type: integer }
      required: [numero, totalEpisodios]
    Episodio:
      type: object
      description: Episódio de um podcast.
      properties:
        numero: { type: integer, example: 1 }
        titulo: { type: string, example: O switch com 40 casos }
        duracaoMin: { type: integer, example: 30 }
        publicadoEm: { type: string, format: date, example: '2025-02-05' }
      required: [titulo]
    EpisodioTemporada:
      type: object
      description: Episódio de uma temporada de série.
      properties:
        numero: { type: integer, example: 1 }
        titulo: { type: string, example: Piloto }
        duracaoMin: { type: integer, example: 42 }
        sinopse: { type: string }
      required: [numero]
    SearchItem:
      allOf:
        - $ref: '#/components/schemas/Midia'
        - type: object
          properties:
            score: { type: number, format: double, description: Relevância (maior = melhor). }
            rank: { type: integer, description: Posição 1-based no ranking. }
    ResultadoBusca:
      type: object
      properties:
        query: { type: string }
        modo: { type: string, enum: [auto, hybrid, vector, fts] }
        usouFallback: { type: boolean, description: True quando caiu para busca léxica (Gemini indisponível). }
        itens:
          type: array
          items: { $ref: '#/components/schemas/SearchItem' }
      required: [query, modo, usouFallback, itens]
    PaginaMidias:
      type: object
      description: >-
        Envelope de toda listagem. O par publicado é `pagina`/`porPagina` — é o
        que a tela usa para dizer "página 2 de 6". Na query, `limite`/`offset`
        continuam aceitos por compatibilidade e são convertidos.
      properties:
        itens:
          type: array
          items: { $ref: '#/components/schemas/Midia' }
        pagina: { type: integer, example: 1 }
        porPagina: { type: integer, example: 20 }
        total: { type: integer, example: 137 }
      required: [itens, pagina, porPagina, total]
...
    Problem:
      type: object
      description: Erro no formato RFC 7807.
      properties:
        type: { type: string, example: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
      required: [title, status]